requirements.txt
asgiref==3.5.2
Babel==2.10.3
Django==4.1.1
django-phonenumber-field==7.0.0
gunicorn==20.1.0
phonenumbers==8.12.55
python-decouple==3.6
pytz==2022.2.1
shortuuid==1.0.9
sqlparse==0.4.2
